+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/interfaces/CanEqual.kt
@@ -1,0 +1,7 @@
+package com.windea.breezeframework.core.interfaces
+
+interface CanEqual {
+	override fun equals(other: Any?): Boolean
+	
+	override fun hashCode(): Int
+}