+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/graph/mermaid/MermaidClassSequenceDsl.kt
@@ -1,3 +1,0 @@
-package com.windea.breezeframework.dsl.graph.mermaid
-
-//DELAY unstable raw api