+++ breeze-dsl/src/test/kotlin/com/windea/breezeframework/dsl/graph/mermaid/MermaidClassDiagramDslKtTest.kt
@@ -1,0 +1,36 @@
+package com.windea.breezeframework.dsl.graph.mermaid
+
+import com.windea.breezeframework.dsl.graph.mermaid.MermaidClassDiagramAnnotationType.*
+import kotlin.test.*
+
+class MermaidClassDiagramDslKtTest {
+	@Test //TESTED OK, @ReplaceWith WORKS
+	fun test1() {
+		//居然真的可以这样写。。。
+		println(mermaidClassDiagram {
+			`class`("Person") {
+				property("name")
+				"Gender"(property("gender"))
+			}
+			`class`("BreezeKnight") {
+				property("weapons")
+				property("magics")
+				protected("Int"(property("memberId")))
+				public(method("tellStoriesAndTales"()))
+				method("helloBreezeFramework"())
+				method("weaponAttack"("weaponName"))
+				method("magicCast"("magicName"))
+				method("dragonDrive"("dragon", "licence"))
+			}
+			`class`("Gender") {
+				annotation(Enumeration)
+				property("Male")
+				property("Female")
+				property("ImmortalMale")
+				property("ImmortalFemale")
+			}
+			//relation("Person", "BreezeKnight", MermaidClassDiagramRelationType.Inheritance)
+			("BreezeKnight" inherits "Person")("Here are dragon knights!") cardinality ("*" to "1")
+		})
+	}
+}