+++ breeze-dsl/src/test/kotlin/com/windea/breezeframework/dsl/graph/mermaid/MermaidPieChartDslKtTest.kt
@@ -12,3 +12,5 @@ class MermaidPieChartDslKtTest {
 		})
 	}
 }
+
+annotation class FooAnnotation(val abc: Array<String> = [])